python/requirements.txt
# Core ML Libraries
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Deep Learning Extensions
torch-geometric>=2.3.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Feature Engineering & Hyperparameter Optimization
optuna>=3.2.0
mlflow>=2.5.0
pydantic>=2.0.0
imbalanced-learn>=0.11.0

# API and Serving
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx>=0.24.0

# Data Processing & Databases
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
redis==5.0.1
neo4j>=5.11.0
pymongo>=4.5.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
click>=8.1.0
tqdm>=4.65.0

# Financial Data
yfinance>=0.2.0
alpha-vantage>=2.3.0
fredapi>=0.5.0

# NLP Processing
spacy>=3.6.0
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0

# Model Serving & Persistence
joblib>=1.3.0
pickle5>=0.0.12
cloudpickle>=2.2.0

# Data Quality & Monitoring
great-expectations>=0.17.0
evidently>=0.4.0

# Additional ML utilities
category-encoders>=2.6.0
feature-engine>=1.6.0
shap>=0.42.0
lime>=0.2.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.0

# Network Analysis (for GNN)
networkx>=3.1.0
dgl>=1.1.0

# Additional dependencies
requests==2.31.0
python-multipart==0.0.6
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0